Refactor: navigator와 screen 파일 폴더 분리
