+++ App.js
@@ -1,12 +1,8 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import RootStack from './screens/RootStack';
-import rootReducer from './redux_modules';
+import RootStack from './navigators/RootStack';
+import store from './redux_modules/store';
 import { Provider } from 'react-redux';
-import { configureStore } from '@reduxjs/toolkit';
-
-const store = configureStore({ reducer: rootReducer });
-console.log(store.getState());
 
 const App = () => {
   return (
