+++ lib/signUpAPI.js
@@ -1,10 +1,9 @@
 import { API } from './API';
-import { useSelector } from 'react-redux';
 
 const url = '/auth/sign-up';
 
 // 회원가입 요청 함수
-export const requestSignUp = async (data) => {
+export const requestSignUp = async data => {
   const postingData = {
     email: data.email,
     password: data.password,
