+++ navigators/FollowTopTab.js
@@ -1,8 +1,8 @@
-// 팔로워/팔로우 리스트를 보여주는 top tab 네비게이터입니다.
+// 팔로워, 팔로우 리스트를 보여주는 top tab 네비게이터입니다.
 
 import React from 'react';
-import FollowerScreen from './FollowerScreen';
-import FollowingScreen from './FollowingScreen';
+import FollowerScreen from '../screens/FollowerScreen';
+import FollowingScreen from '../screens/FollowingScreen';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 
 const Tab = createMaterialTopTabNavigator();
